requirements.txt
requests
beautifulsoup4
lxml
colorama
PyGithub
